internal/sandbox/infra: make this build on FreeBSD
This simply moves SetupShm to infra_linux.go since this is not needed on
FreeBSD and the MS_* constants are Linux-specific.
